+++ src/repositories/mssql.repository.ts
@@ -1,17 +1,10 @@
+import { Database } from "../models/database.model";
+import { DatabaseTable } from "../models/database-table.model";
+
 const windowsSql = () => require('mssql/msnodesqlv8');
 const allSql = require('mssql');
 
-export interface MssqlDb {
-  TABLE_NAME: string;
-  C: MssqlColumn[];
-}
 
-export interface MssqlColumn {
-  COLUMN_NAME: string;
-  REFERENCE_COLUMN: string;
-  REFERENCE_TO_TABLE: string;
-  FOREIGN_KEY: string;
-}
 
 export const getMssqlDbSchema = async (
   username: string,
@@ -19,11 +12,14 @@ export const getMssqlDbSchema = async (
   server: string,
   database: string,
   trusted: boolean
-): Promise<MssqlDb[]> => {
-  let db = [] as MssqlDb[];
+): Promise<Database> => {
+  let db: Database = {
+    tables: [],
+    errors: []
+  } as Database;
   let sql: any;
   try {
-    if(trusted) {
+    if (trusted) {
       sql = windowsSql();
     } else {
       sql = allSql;
@@ -31,78 +27,62 @@ export const getMssqlDbSchema = async (
     const config = {
       user: username,
       password: password,
-      server: server === "localhost" ? server : "localhost:1433",
+      server: server,
       database: database,
-      driver: trusted ? "{SQL Server Native Client 11.0}": "",
+      driver: trusted ? "msnodesqlv8" : "",
       options: {
         trustedConnection: trusted
       }
     };
     await sql.connect(config);
-    const result = await sql.query`
-  SELECT (SELECT 
-      T.TABLE_NAME,
-      C.COLUMN_NAME,
-      (SELECT TOP 1
-      TAB2.NAME
-      FROM SYS.FOREIGN_KEY_COLUMNS fkc
-      INNER JOIN SYS.OBJECTS obj
-          ON OBJ.OBJECT_ID = fkc.constraint_object_id
-      INNER JOIN SYS.TABLES tab1
-          ON tab1.OBJECT_ID = fkc.parent_object_id
-      INNER JOIN sys.schemas sch
-          ON tab1.schema_id = sch.schema_id
-      INNER JOIN sys.columns col1
-          ON col1.column_id = parent_column_id AND col1.object_id = tab1.object_id
-      INNER JOIN sys.tables tab2
-          ON tab2.OBJECT_ID = fkc.referenced_object_id
-      INNER JOIN sys.columns col2
-          ON col2.column_id = referenced_column_id AND col2.object_id = tab2.object_id
-      WHERE col1.name = C.COLUMN_NAME AND C.TABLE_NAME != TAB2.NAME) as REFERENCE_TO_TABLE,
-      (SELECT TOP 1
-      col2.name AS [referenced_column]
-      FROM SYS.FOREIGN_KEY_COLUMNS fkc
-      INNER JOIN SYS.OBJECTS obj
-          ON OBJ.OBJECT_ID = fkc.constraint_object_id
-      INNER JOIN SYS.TABLES tab1
-          ON tab1.OBJECT_ID = fkc.parent_object_id
-      INNER JOIN sys.schemas sch
-          ON tab1.schema_id = sch.schema_id
-      INNER JOIN sys.columns col1
-          ON col1.column_id = parent_column_id AND col1.object_id = tab1.object_id
-      INNER JOIN sys.tables tab2
-          ON tab2.OBJECT_ID = fkc.referenced_object_id
-      INNER JOIN sys.columns col2
-          ON col2.column_id = referenced_column_id AND col2.object_id = tab2.object_id
-      WHERE col1.name = C.COLUMN_NAME AND C.TABLE_NAME != TAB2.NAME) as REFERENCE_COLUMN,
-       (SELECT TOP 1
-      OBJ.NAME
-      FROM SYS.FOREIGN_KEY_COLUMNS fkc
-      INNER JOIN SYS.OBJECTS obj
-          ON OBJ.OBJECT_ID = fkc.constraint_object_id
-      INNER JOIN SYS.TABLES tab1
-          ON tab1.OBJECT_ID = fkc.parent_object_id
-      INNER JOIN sys.schemas sch
-          ON tab1.schema_id = sch.schema_id
-      INNER JOIN sys.columns col1
-          ON col1.column_id = parent_column_id AND col1.object_id = tab1.object_id
-      INNER JOIN sys.tables tab2
-          ON tab2.OBJECT_ID = fkc.referenced_object_id
-      INNER JOIN sys.columns col2
-          ON col2.column_id = referenced_column_id AND col2.object_id = tab2.object_id
-      WHERE col1.name = C.COLUMN_NAME AND C.TABLE_NAME != TAB2.NAME) as FOREIGN_KEY
-  
-      FROM INFORMATION_SCHEMA.TABLES T
-      INNER JOIN INFORMATION_SCHEMA.COLUMNS C ON C.TABLE_NAME = T.TABLE_NAME
-      WHERE T.TABLE_TYPE='BASE TABLE' FOR JSON AUTO) AS 'JSONRESULT'`;
 
-    db = JSON.parse(result.recordset[0]['JSONRESULT']) as MssqlDb[];
+    let dbEntry = await getFromSysTables(sql);
+
+    if (!dbEntry) {
+      dbEntry = await sql.query(informationsSchemaQuery);
+    }
+
+    db.tables = toTables(dbEntry);
   } catch (err) {
-    // ... error checks
+    db.errors.push('Error getting data: ' + err.message);
   } finally {
     sql.close();
   }
   return db;
+};
+
+const toTables = (dbResult): DatabaseTable[] => {
+  const result: DatabaseTable[] = [];
+  for (let index = 0; index < dbResult.recordset.length; index++) {
+    const element = dbResult.recordset[index];
+    const columnToAdd = {
+      Name: element.COLUMN_NAME,
+      ReferenceColumn: element.REFERENCE_COLUMN,
+      ReferenceTable: element.REFERENCE_TO_TABLE,
+      ForeignKey: element.FOREIGN_KEY,
+    };
+    const existing = result.find(t => t.Name === element.TABLE_NAME);
+
+    if (existing) {
+      existing.Columns.push(columnToAdd);
+    } else {
+      result.push({
+        Name: element.TABLE_NAME,
+        Columns: [
+          columnToAdd
+        ]
+      });
+    }
+  }
+  return result;
+};
+
+const getFromSysTables = async (sql) => {
+  try {
+    return await sql.query(sysTableQuery);
+  } catch (error) {
+
+  }
 };
 
 export const listDatabases = async (
@@ -123,3 +103,113 @@ export const listDatabases = async (
   }
   return databases;
 };
+
+const sysTableQuery = `
+SELECT
+    T.name as 'TABLE_NAME',
+    Columns.name as 'COLUMN_NAME',
+    (SELECT TOP 1
+        TAB2.NAME
+    FROM SYS.FOREIGN_KEY_COLUMNS fkc
+        INNER JOIN SYS.OBJECTS obj
+        ON OBJ.OBJECT_ID = fkc.constraint_object_id
+        INNER JOIN SYS.TABLES tab1
+        ON tab1.OBJECT_ID = fkc.parent_object_id
+        INNER JOIN sys.schemas sch
+        ON tab1.schema_id = sch.schema_id
+        INNER JOIN sys.columns col1
+        ON col1.column_id = parent_column_id AND col1.object_id = tab1.object_id
+        INNER JOIN sys.tables tab2
+        ON tab2.OBJECT_ID = fkc.referenced_object_id
+        INNER JOIN sys.columns col2
+        ON col2.column_id = referenced_column_id AND col2.object_id = tab2.object_id
+    WHERE (col1.name != col2.NAME OR tab1.name != tab2.name) AND tab1.name = t.name AND col1.name = Columns.name) as REFERENCE_TO_TABLE,
+    (SELECT TOP 1
+        col2.name AS [referenced_column]
+    FROM SYS.FOREIGN_KEY_COLUMNS fkc
+        INNER JOIN SYS.OBJECTS obj
+        ON OBJ.OBJECT_ID = fkc.constraint_object_id
+        INNER JOIN SYS.TABLES tab1
+        ON tab1.OBJECT_ID = fkc.parent_object_id
+        INNER JOIN sys.schemas sch
+        ON tab1.schema_id = sch.schema_id
+        INNER JOIN sys.columns col1
+        ON col1.column_id = parent_column_id AND col1.object_id = tab1.object_id
+        INNER JOIN sys.tables tab2
+        ON tab2.OBJECT_ID = fkc.referenced_object_id
+        INNER JOIN sys.columns col2
+        ON col2.column_id = referenced_column_id AND col2.object_id = tab2.object_id
+    WHERE (col1.name != col2.NAME OR tab1.name != tab2.name) AND tab1.name = t.name AND col1.name = Columns.name) as REFERENCE_COLUMN,
+    (SELECT TOP 1
+        OBJ.NAME
+    FROM SYS.FOREIGN_KEY_COLUMNS fkc
+        INNER JOIN SYS.OBJECTS obj
+        ON OBJ.OBJECT_ID = fkc.constraint_object_id
+        INNER JOIN SYS.TABLES tab1
+        ON tab1.OBJECT_ID = fkc.parent_object_id
+        INNER JOIN sys.schemas sch
+        ON tab1.schema_id = sch.schema_id
+        INNER JOIN sys.columns col1
+        ON col1.column_id = parent_column_id AND col1.object_id = tab1.object_id
+        INNER JOIN sys.tables tab2
+        ON tab2.OBJECT_ID = fkc.referenced_object_id
+        INNER JOIN sys.columns col2
+        ON col2.column_id = referenced_column_id AND col2.object_id = tab2.object_id
+    WHERE (col1.name != col2.NAME OR tab1.name != tab2.name) AND tab1.name = t.name AND col1.name = Columns.name) as FOREIGN_KEY
+
+FROM SYS.TABLES T
+    INNER JOIN SYS.COLUMNS Columns ON Columns.object_id = T.object_id
+`;
+
+const informationsSchemaQuery = `
+SELECT
+        T.TABLE_NAME,
+        Columns.COLUMN_NAME,
+        (SELECT
+            KCU2.TABLE_NAME
+        FROM INFORMATION_SCHEMA.REFERENTIAL_CONSTRAINTS RC
+            LEFT JOIN INFORMATION_SCHEMA.KEY_COLUMN_USAGE KCU1 ON 
+        KCU1.CONSTRAINT_CATALOG = RC.CONSTRAINT_CATALOG AND
+                KCU1.CONSTRAINT_SCHEMA = RC.CONSTRAINT_SCHEMA AND
+                KCU1.CONSTRAINT_NAME = RC.CONSTRAINT_NAME
+            LEFT JOIN INFORMATION_SCHEMA.KEY_COLUMN_USAGE KCU2 ON 
+        KCU2.CONSTRAINT_CATALOG = RC.UNIQUE_CONSTRAINT_CATALOG AND
+                KCU2.CONSTRAINT_SCHEMA = RC.UNIQUE_CONSTRAINT_SCHEMA AND
+                KCU2.CONSTRAINT_NAME = RC.UNIQUE_CONSTRAINT_NAME
+        WHERE KCU1.ORDINAL_POSITION = KCU2.ORDINAL_POSITION AND
+            KCU1.TABLE_NAME = T.TABLE_NAME AND KCU1.COLUMN_NAME = Columns.COLUMN_NAME) as REFERENCE_TO_TABLE,
+
+        (SELECT
+            KCU2.COLUMN_NAME
+        FROM INFORMATION_SCHEMA.REFERENTIAL_CONSTRAINTS RC
+            LEFT JOIN INFORMATION_SCHEMA.KEY_COLUMN_USAGE KCU1 ON 
+        KCU1.CONSTRAINT_CATALOG = RC.CONSTRAINT_CATALOG AND
+                KCU1.CONSTRAINT_SCHEMA = RC.CONSTRAINT_SCHEMA AND
+                KCU1.CONSTRAINT_NAME = RC.CONSTRAINT_NAME
+            LEFT JOIN INFORMATION_SCHEMA.KEY_COLUMN_USAGE KCU2 ON 
+        KCU2.CONSTRAINT_CATALOG = RC.UNIQUE_CONSTRAINT_CATALOG AND
+                KCU2.CONSTRAINT_SCHEMA = RC.UNIQUE_CONSTRAINT_SCHEMA AND
+                KCU2.CONSTRAINT_NAME = RC.UNIQUE_CONSTRAINT_NAME
+        WHERE KCU1.ORDINAL_POSITION = KCU2.ORDINAL_POSITION AND
+            KCU1.TABLE_NAME = T.TABLE_NAME AND KCU1.COLUMN_NAME = Columns.COLUMN_NAME) as REFERENCE_COLUMN,
+
+
+        (SELECT KCU1.CONSTRAINT_NAME AS 'FOREIGN_KEY'
+        FROM INFORMATION_SCHEMA.REFERENTIAL_CONSTRAINTS RC
+            LEFT JOIN INFORMATION_SCHEMA.KEY_COLUMN_USAGE KCU1 ON 
+        KCU1.CONSTRAINT_CATALOG = RC.CONSTRAINT_CATALOG AND
+                KCU1.CONSTRAINT_SCHEMA = RC.CONSTRAINT_SCHEMA AND
+                KCU1.CONSTRAINT_NAME = RC.CONSTRAINT_NAME
+            LEFT JOIN INFORMATION_SCHEMA.KEY_COLUMN_USAGE KCU2 ON 
+        KCU2.CONSTRAINT_CATALOG = RC.UNIQUE_CONSTRAINT_CATALOG AND
+                KCU2.CONSTRAINT_SCHEMA = RC.UNIQUE_CONSTRAINT_SCHEMA AND
+                KCU2.CONSTRAINT_NAME = RC.UNIQUE_CONSTRAINT_NAME
+        WHERE KCU1.ORDINAL_POSITION = KCU2.ORDINAL_POSITION AND
+            KCU1.TABLE_NAME = T.TABLE_NAME AND KCU1.COLUMN_NAME = Columns.COLUMN_NAME
+) as FOREIGN_KEY
+
+    FROM INFORMATION_SCHEMA.TABLES T
+        INNER JOIN INFORMATION_SCHEMA.COLUMNS Columns ON Columns.TABLE_NAME = T.TABLE_NAME
+ WHERE T.TABLE_TYPE='BASE TABLE'
+
+`;