+++ CHANGELOG.md
@@ -3,5 +3,14 @@ All notable changes to the "Schema Visualization" extension will be documented i
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
-## [Unreleased]
+## [0.3.0]
+- Added fallback way of getting the schema
+- Removed JSON AUTO dependency
+- Added error handling and ability to show errors to the user.
+
+## [0.2.0]
+- Added support for windows auth
+
+## [0.1.0]
 - Initial release
+